start some refactoring
